+++ bin/qtextasdata
@@ -1,0 +1,1 @@
+/Users/harel/dev/harelba/refactor-q/qtextasdata
